+++ tests/integration/test_transaction.py
@@ -789,10 +789,10 @@ class TestTransaction(unittest.TestCase):
         self.assertTrue(int(apple_pay_details.expiration_year) > 0)
         self.assertNotEqual(None, apple_pay_details.cardholder_name)
 
-    def test_sale_with_fake_android_pay_nonce(self):
+    def test_sale_with_fake_android_pay_proxy_card_nonce(self):
         result = Transaction.sale({
             "amount": "10.00",
-            "payment_method_nonce": Nonces.AndroidPayCard
+            "payment_method_nonce": Nonces.AndroidPayCardDiscover
         })
 
         self.assertTrue(result.is_success)
@@ -801,6 +801,21 @@ class TestTransaction(unittest.TestCase):
         android_pay_card_details = result.transaction.android_pay_card_details
         self.assertNotEqual(None, android_pay_card_details)
         self.assertEqual(CreditCard.CardType.Discover, android_pay_card_details.card_type)
+        self.assertTrue(int(android_pay_card_details.expiration_month) > 0)
+        self.assertTrue(int(android_pay_card_details.expiration_year) > 0)
+
+    def test_sale_with_fake_android_pay_network_token_nonce(self):
+        result = Transaction.sale({
+            "amount": "10.00",
+            "payment_method_nonce": Nonces.AndroidPayCardMasterCard
+        })
+
+        self.assertTrue(result.is_success)
+        self.assertEqual(result.transaction.amount, 10.00)
+        self.assertEqual(result.transaction.payment_instrument_type, PaymentInstrumentType.AndroidPayCard)
+        android_pay_card_details = result.transaction.android_pay_card_details
+        self.assertNotEqual(None, android_pay_card_details)
+        self.assertEqual(CreditCard.CardType.MasterCard, android_pay_card_details.card_type)
         self.assertTrue(int(android_pay_card_details.expiration_month) > 0)
         self.assertTrue(int(android_pay_card_details.expiration_year) > 0)
 
@@ -2072,6 +2087,106 @@ class TestTransaction(unittest.TestCase):
             ErrorCodes.Transaction.ThreeDSecureTransactionDataDoesntMatchVerify,
             result.errors.for_object("transaction").on("three_d_secure_token")[0].code
         )
+
+    def test_sale_returns_a_successful_result_with_amex_rewards_response(self):
+        result = Transaction.sale({
+            "amount": TransactionAmounts.Authorize,
+            "credit_card": {
+                "number": "371449635392376",
+                "expiration_date": "05/2009"
+            },
+            "options" : {
+                "submit_for_settlement" : True,
+                "amex_rewards" : {
+                    "request_id" : "ABC123",
+                    "points" : "100",
+                    "currency_amount" : "1.00",
+                    "currency_iso_code" : "USD"
+                }
+            }
+        })
+
+        self.assertTrue(result.is_success)
+        transaction = result.transaction
+        self.assertEquals(Transaction.Type.Sale, transaction.type)
+        self.assertEquals(Transaction.Status.SubmittedForSettlement, transaction.status)
+        self.assertEquals("success", transaction.amex_rewards_response)
+
+    def test_sale_returns_an_unsuccessful_result_with_amex_rewards_response(self):
+        result = Transaction.sale({
+            "amount": TransactionAmounts.Authorize,
+            "credit_card": {
+                "number": "371449635392376",
+                "expiration_date": "05/2009"
+            },
+            "options" : {
+                "submit_for_settlement" : True,
+                "amex_rewards" : {
+                    "request_id" : "CARD_INELIGIBLE",
+                    "points" : "100",
+                    "currency_amount" : "1.00",
+                    "currency_iso_code" : "USD"
+                }
+            }
+        })
+
+        self.assertTrue(result.is_success)
+        transaction = result.transaction
+        self.assertEquals(Transaction.Type.Sale, transaction.type)
+        self.assertEquals(Transaction.Status.SubmittedForSettlement, transaction.status)
+        self.assertEquals("RDM2002 Card is not eligible for redemption", transaction.amex_rewards_response)
+
+    def test_submit_for_settlement_returns_a_successful_result_with_amex_rewards_response(self):
+        result = Transaction.sale({
+            "amount": TransactionAmounts.Authorize,
+            "credit_card": {
+                "number": "371449635392376",
+                "expiration_date": "05/2009"
+            },
+            "options" : {
+                "amex_rewards" : {
+                    "request_id" : "ABC123",
+                    "points" : "100",
+                    "currency_amount" : "1.00",
+                    "currency_iso_code" : "USD"
+                }
+            }
+        })
+
+        self.assertTrue(result.is_success)
+        transaction = result.transaction
+        self.assertEquals(Transaction.Type.Sale, transaction.type)
+        self.assertEquals(Transaction.Status.Authorized, transaction.status)
+
+        submitted_transaction = Transaction.submit_for_settlement(transaction.id).transaction
+        self.assertEquals("success", submitted_transaction.amex_rewards_response)
+        self.assertEquals(Transaction.Status.SubmittedForSettlement, submitted_transaction.status)
+
+    def test_submit_for_settlement_returns_an_unsuccessful_result_with_amex_rewards_response(self):
+        result = Transaction.sale({
+            "amount": TransactionAmounts.Authorize,
+            "credit_card": {
+                "number": "371449635392376",
+                "expiration_date": "05/2009"
+            },
+            "options" : {
+                "amex_rewards" : {
+                    "request_id" : "CARD_INELIGIBLE",
+                    "points" : "100",
+                    "currency_amount" : "1.00",
+                    "currency_iso_code" : "USD"
+                }
+            }
+        })
+
+        self.assertTrue(result.is_success)
+        transaction = result.transaction
+        self.assertEquals(Transaction.Type.Sale, transaction.type)
+        self.assertEquals(Transaction.Status.Authorized, transaction.status)
+
+        submitted_transaction = Transaction.submit_for_settlement(transaction.id).transaction
+        self.assertEquals("RDM2002 Card is not eligible for redemption", submitted_transaction.amex_rewards_response)
+        self.assertEquals(Transaction.Status.SubmittedForSettlement, submitted_transaction.status)
 
     def test_find_exposes_disputes(self):
         transaction = Transaction.find("disputedtransaction")